% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitLN.R
\name{unitLN}
\alias{unitLN}
\title{Small area inference for the unit-level lognormal model}
\usage{
unitLN(yspos, Xs, Xpop, areafacpop, areafacsamp, sampindex)
}
\arguments{
\item{yspos}{a numeric vector with the positive (not log transformed) response variables}

\item{Xs}{the matrix of covariates for sampled elements that does not contain an intercept}

\item{Xpop}{the matrix of covariates for the full population that does not contain an intercept}

\item{areafacpop}{a vector of area labels for the full population}

\item{areafacsamp}{a vector of area samples for the full population}

\item{sampindex}{the vector of sampled index values}
}
\value{
a list with predictions and MSE estimates
}
\description{
Small area inference for the unit-level lognormal model
}
\examples{
beta0 <- log(0.5)
beta1 <- 1
beta2 <- 2
D <- 60
Nis <- c(100, 200, 500)
Nis <- rep(Nis, each = D/length(Nis))
N <- sum(Nis)
x1 <- rnorm(N)
x2 <- rnorm(N)
areafacpop <- rep(1:D, Nis)
sigma2b <- 0.5
sigma2e <- 1
bi <- rnorm(D, mean = 0, sd = sqrt(sigma2b))
ei <- rnorm(N, mean = 0, sd = sqrt(sigma2e))
names(bi) <- as.character(1:D)
y <- exp(beta0 + beta1*x1 + beta2*x2 + bi[as.character(areafacpop)] + ei)
samplist <- sapply(1:D, function(i){sample( (1:N)[areafacpop == i],
size = 0.1*Nis[i], replace = FALSE)})
sampindex <- unlist(samplist)
areafacsamp <- areafacpop[sampindex]
ys <- y[sampindex]
Xpop <- cbind(x1, x2)
Xs <- Xpop[sampindex,]
popindex <- 1:N
unitLN(ys, Xs, Xpop,  areafacpop, areafacsamp, sampindex)
}
